boundary handling
